refactor: restructure SMTC module by creating a new test directory
- Deleted existing SMTC related files and created a new directory for testing.
- Moved and refactored components, hooks, and utility functions into the new `smtc-test` directory.
- Introduced new `MarqueeText` and `SmtcStep` components for better modularity.
- Implemented hooks for managing state and color extraction.
- Updated tests to reflect the new structure and ensure functionality remains intact.
