+++ src/renderer/components/components/DynamicIslandGuidePages/smtc-test/components/MarqueeText.tsx
