+++ src/renderer/components/components/DynamicIslandGuidePages/smtc-test/components/SmtcStep.tsx
