+++ src/renderer/components/components/DynamicIslandGuidePages/smtc-test/hooks/useSmtcAccentColor.ts
