+++ src/renderer/components/components/DynamicIslandGuidePages/smtc-test/hooks/useSmtcTest.ts
