+++ src/renderer/components/components/DynamicIslandGuidePages/smtc-test/index.ts
