+++ src/renderer/components/components/DynamicIslandGuidePages/smtc-test/types/index.ts
