+++ src/renderer/components/components/DynamicIslandGuidePages/smtc-test/types/media.ts
