+++ src/renderer/components/components/DynamicIslandGuidePages/smtc-test/types/props.ts
