+++ src/renderer/components/components/DynamicIslandGuidePages/smtc-test/utils/smtcActions.ts
