+++ src/renderer/components/components/DynamicIslandGuidePages/smtc-test/utils/smtcStore.ts
