+++ src/renderer/components/components/DynamicIslandGuidePages/smtc-test/utils/smtcUtils.ts
